pyproject.toml
[project]
name = "qinglong-mcp-server"
version = "0.1.3"
description = "MCP server for Qinglong Panel task management"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["mcp", "qinglong", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests",
    "python-dotenv"
]

[project.scripts]
qinglong-mcp-server = "server:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["server.py"]

[tool.hatch.build.targets.sdist]
only-include = ["server.py", "README.md", "pyproject.toml", ".env.example"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
